Clean up, fix tests, update deps
