+++ .eslintignore
@@ -2,3 +2,4 @@
 coverage
 site/themes/*
 convert
+node_modules
