+++ .eslintrc
@@ -2,13 +2,16 @@
   "extends": [
     "airbnb",
     "plugin:ava/recommended",
+    "plugin:jsdoc/recommended",
     "plugin:node/recommended",
+    "plugin:optimize-regex/all",
     "plugin:ramda/recommended",
     "plugin:security/recommended",
     "plugin:unicorn/recommended"
   ],
   "plugins": [
     "ava",
+    "jsdoc",
     "no-inferred-method-name",
     "node",
     "optimize-regex",
@@ -28,9 +31,14 @@
     "expect": true
   },
   "parserOptions": {
-    "ecmaVersion": 2018
+    "ecmaVersion": 2020
   },
   "rules": {
+    "ava/no-only-test": 0,
+    "camelcase": 0,
+    "class-methods-use-this": 0,
+    "consistent-return": ["warn", { "treatUndefinedAsUnspecified": false }],
+    "global-require": 0,
     "import/no-dynamic-require": 0,
     "import/no-extraneous-dependencies": [
       "error", {
@@ -39,20 +47,22 @@
         "peerDependencies": false
       }
     ],
-    "node/no-unpublished-require": 0,
-    "camelcase": 0,
-    "global-require": 0,
     "max-len": 0,
     "no-empty": ["error", { "allowEmptyCatch": true }],
+    "no-inferred-method-name/no-inferred-method-name": "error",
     "no-param-reassign": 0,
     "no-plusplus": 0,
     "no-restricted-syntax": 0,
     "no-underscore-dangle": 0,
     "no-unused-vars": ["error", { "varsIgnorePattern": "^_", "argsIgnorePattern": "^_", "caughtErrors": "none" }],
-    "security/detect-object-injection": 0,
+    "node/no-unpublished-require": 0,
+    "object-curly-newline": 0,
     "optimize-regex/optimize-regex": "warn",
+    "security/detect-non-literal-fs-filename": 0,
+    "security/detect-non-literal-require": 0,
+    "security/detect-object-injection": 0,
+    "unicorn/catch-error-name": 0,
     "unicorn/filename-case": 0,
-    "unicorn/prevent-abbreviations": 0,
-    "ava/no-only-test": 0
+    "unicorn/prevent-abbreviations": 0
   }
 }
