+++ docs/index.html
@@ -2,9 +2,9 @@
 <html lang="en">
 <head>
   <meta charset="UTF-8">
-  <title>uttori-wiki - Uttori Wiki is a wiki functionality module for Uttori.</title>
+  <title>@uttori/wiki - Uttori Wiki is a Wiki using Uttori Compoents.</title>
   <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
-  <meta name="description" content="Uttori Wiki is a wiki functionality module for Uttori.">
+  <meta name="description" content="@uttori/wiki - Uttori Wiki is a Wiki using Uttori Compoents">
   <meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
   <link rel="stylesheet" href="//unpkg.com/docsify/lib/themes/vue.css">
 </head>
@@ -12,7 +12,7 @@
   <div id="app"></div>
   <script>
     window.$docsify = {
-      name: 'uttori-wiki',
+      name: '@uttori/wiki',
       repo: 'uttori/uttori-wiki'
     }
   </script>
