+++ package.json
@@ -1,59 +1,74 @@
 {
-  "name": "uttori-wiki",
-  "version": "2.4.1-pre",
+  "name": "@uttori/wiki",
+  "version": "3.0.0",
   "description": "Uttori Wiki is a wiki functionality module for Uttori.",
   "author": "Matthew Callis <matthew.callis@gmail.com>",
   "contributors": [],
   "license": "MIT",
-  "repository": "uttori/uttori-wiki",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/uttori/uttori-wiki.git"
+  },
+  "bugs": {
+    "url": "https://github.com/uttori/uttori-wiki/issues"
+  },
   "homepage": "https://github.com/uttori/uttori-wiki",
   "keywords": [
     "uttori",
     "wiki"
   ],
   "engines": {
-    "node": ">= 10.15.3"
+    "node": ">= 12"
   },
   "dependencies": {
     "debug": "^4.1.1",
     "ramda": "^0.27.0",
     "uttori-document": "^1.4.1",
-    "uttori-utilities": "^2.1.1"
+    "uttori-utilities": "^2.3.0"
   },
   "optionalDependencies": {},
   "devDependencies": {
-    "ava": "^3.2.0",
+    "@uttori/search-provider-lunr": "^3.2.0",
+    "@uttori/storage-provider-json-memory": "^3.2.0",
+    "ava": "^3.9.0",
     "body-parser": "^1.19.0",
-    "coveralls": "^3.0.9",
-    "docsify-cli": "^4.4.0",
-    "ejs": "^3.0.1",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-plugin-ava": "^10.0.1",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "coveralls": "^3.1.0",
+    "docsify-cli": "^4.4.1",
+    "ejs": "^3.1.3",
+    "eslint": "7.2.0",
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-plugin-ava": "^10.3.1",
+    "eslint-plugin-import": "^2.21.2",
+    "eslint-plugin-jsdoc": "^27.1.2",
+    "eslint-plugin-jsx-a11y": "^6.3.1",
     "eslint-plugin-no-inferred-method-name": "^2.0.0",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-optimize-regex": "^1.1.7",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-optimize-regex": "^1.2.0",
     "eslint-plugin-ramda": "^2.5.1",
-    "eslint-plugin-react": "^7.18.3",
-    "eslint-plugin-react-hooks": "^2.3.0",
+    "eslint-plugin-react": "^7.20.0",
+    "eslint-plugin-react-hooks": "^4.0.4",
     "eslint-plugin-security": "^1.4.0",
-    "eslint-plugin-unicorn": "^16.0.0",
-    "eslint-plugin-xss": "^0.1.9",
+    "eslint-plugin-unicorn": "^20.1.0",
+    "eslint-plugin-xss": "^0.1.10",
     "express": "^4.17.1",
     "express-ejs-layouts": "^2.5.0",
-    "jsdoc-to-markdown": "^5.0.3",
-    "markdown-it": "^10.0.0",
-    "nyc": "^15.0.0",
-    "sinon": "^8.1.1",
-    "supertest": "^4.0.2",
-    "uttori-storage-provider-json-memory": "^1.0.1"
+    "jsdoc-to-markdown": "^6.0.1",
+    "jsdoc": "^3.6.4",
+    "markdown-it": "^11.0.0",
+    "nyc": "^15.1.0",
+    "pre-commit": "^1.2.2",
+    "release-it": "^13.6.3",
+    "sinon": "^9.0.2",
+    "tsd-jsdoc": "^2.5.0",
+    "supertest": "^4.0.2"
   },
   "files": [
-    "src/*"
+    "src/*",
+    "types/*"
   ],
   "main": "src/index.js",
+  "types": "types/index.d.ts",
+  "typings": "types/index.d.ts",
   "ava": {
     "files": [
       "test/**/*.test.js"
@@ -76,14 +91,23 @@
     "exclude": []
   },
   "scripts": {
-    "docs": "jsdoc2md --private --example-lang js --template README.hbs src/index.js > README.md; cp README.md docs/README.md",
+    "docs": "jsdoc2md --private --example-lang js --template rm.hbs src/index.js > README.md; jsdoc2md --private --example-lang js src/index.js > docs/wiki.md;",
     "docs-serve": "docsify init ./docs && docsify serve ./docs",
     "lint": "eslint .",
+    "make-types": "jsdoc -t node_modules/tsd-jsdoc/dist -r ./src --destination ./; mv types.d.ts types/index.d.ts",
+    "release": "release-it",
     "report": "nyc report --reporter=html",
     "pretest": "killall -SIGINT uttori-wiki-test-server && node test/_helpers/server.js uttori-wiki-test-server &",
     "posttest": "killall -SIGINT uttori-wiki-test-server || true",
     "test": "NODE_ENV=test nyc ava",
-    "test-debug": "DEBUG=Uttori* NODE_ENV=test ava --serial",
+    "test-debug": "DEBUG=Uttori* NODE_ENV=test ava --serial --concurrency=1",
     "validate": "npm ls"
-  }
+  },
+  "directories": {
+    "test": "test"
+  },
+  "pre-commit": [
+    "lint",
+    "test"
+  ]
 }
