+++ src/middleware.js
@@ -1,0 +1,6 @@
+module.exports = function (options) {
+  return function (req, res, next) {
+    // Implement the middleware function based on the options object
+    next();
+  };
+};
