+++ test/_helpers/server.js
@@ -1,14 +1,12 @@
-/* eslint-disable no-console */
-// Server
 const express = require('express');
 const ejs = require('ejs');
 const layouts = require('express-ejs-layouts');
 const bodyParser = require('body-parser');
 const path = require('path');
 
-const StorageProvider = require('uttori-storage-provider-json-memory');
-const SearchProvider = require('./../__stubs/SearchProvider.js');
-const defaultConfig = require('../../src/config.default.js');
+const { Plugin: StorageProviderJSON } = require('@uttori/storage-provider-json-memory');
+const { Plugin: SearchProviderLunr } = require('@uttori/search-provider-lunr');
+const defaultConfig = require('../../src/config');
 
 const config = {
   ...defaultConfig,
@@ -27,12 +25,10 @@ const config = {
   public_dir: 'test/site/themes/default/public',
   use_delete_key: true,
   delete_key: 'test-key',
-
-  // Providers
-  StorageProvider,
-  storageProviderConfig: {},
-  SearchProvider,
-  searchProviderConfig: {},
+  plugins: [
+    StorageProviderJSON,
+    SearchProviderLunr,
+  ],
 };
 
 const serverSetup = () => {
@@ -55,6 +51,7 @@ const serverSetup = () => {
   server.use(bodyParser.urlencoded({ limit: '50mb', extended: false }));
 
   if (process.argv[2] && process.argv[2] !== 'undefined') {
+    // eslint-disable-next-line no-console
     console.log('Setting process.title:', typeof process.argv[2], process.argv[2]);
     // eslint-disable-next-line prefer-destructuring
     process.title = process.argv[2];
@@ -62,6 +59,7 @@ const serverSetup = () => {
 
   // Is this a require()?
   if (require.main === module) {
+    // eslint-disable-next-line no-console
     console.log('Starting test server...');
     server.listen(server.get('port'), server.get('ip'));
   }
@@ -69,7 +67,9 @@ const serverSetup = () => {
   return server;
 };
 
-const seed = (storageProvider) => {
+const seed = async (uttori) => {
+  const response = { set: () => {}, render: () => {}, redirect: () => {} };
+  const next = function next() {};
   const demoTitle = {
     title: 'Demo Title Beta',
     slug: 'demo-title',
@@ -77,9 +77,11 @@ const seed = (storageProvider) => {
     html: '',
     updateDate: 1459310452002,
     createDate: 1459310452002,
-    tags: ['Demo Tag', 'Cool'],
+    tags: 'Demo Tag,Cool',
   };
-  storageProvider.add(demoTitle);
+  // await uttori.hooks.dispatch('storage-add', demoTitle, uttori);
+  // await uttori.hooks.dispatch('search-add', [demoTitle], uttori);
+  await uttori.saveValid({ params: {}, body: demoTitle }, response, next);
 
   const demoTitleHistory = {
     title: 'Demo Title',
@@ -88,9 +90,11 @@ const seed = (storageProvider) => {
     html: '',
     updateDate: 1500000000000,
     createDate: 1500000000000,
-    tags: ['Demo Tag', 'Cool'],
+    tags: 'Demo Tag,Cool',
   };
-  storageProvider.update(demoTitleHistory, 'demo-title');
+  // await uttori.hooks.dispatch('storage-update', { document: demoTitleHistory, originalSlug: 'demo-title' }, uttori);
+  // await uttori.hooks.dispatch('search-update', [{ document: demoTitleHistory, originalSlug: 'demo-title' }], uttori);
+  await uttori.saveValid({ params: {}, body: demoTitleHistory }, response, next);
 
   const exampleTitle = {
     title: 'Example Title',
@@ -99,9 +103,11 @@ const seed = (storageProvider) => {
     html: '',
     updateDate: 1459310452001,
     createDate: 1459310452001,
-    tags: ['Example Tag', 'example'],
+    tags: 'Example Tag,example',
   };
-  storageProvider.add(exampleTitle);
+  // await uttori.hooks.dispatch('storage-add', exampleTitle, uttori);
+  // await uttori.hooks.dispatch('search-add', [exampleTitle], uttori);
+  await uttori.saveValid({ params: {}, body: exampleTitle }, response, next);
 
   const fakeTitle = {
     title: 'Fake Title',
@@ -110,20 +116,24 @@ const seed = (storageProvider) => {
     html: '',
     updateDate: 1459310452000,
     createDate: 1459310452000,
-    tags: ['Fake Tag', 'Cool'],
+    tags: 'Fake Tag,Cool',
   };
-  storageProvider.add(fakeTitle);
+  // await uttori.hooks.dispatch('storage-add', fakeTitle, uttori);
+  // await uttori.hooks.dispatch('search-add', [fakeTitle], uttori);
+  await uttori.saveValid({ params: {}, body: fakeTitle }, response, next);
 
   const homePage = {
     content: '## Home Page',
-    createDate: null,
+    createDate: undefined,
     html: '',
     slug: 'home-page',
-    tags: [],
+    tags: '',
     title: 'Home Page',
     updateDate: 1512921841841,
   };
-  storageProvider.add(homePage);
+  // await uttori.hooks.dispatch('storage-add', homePage, uttori);
+  // await uttori.hooks.dispatch('search-add', [homePage], uttori);
+  await uttori.saveValid({ params: {}, body: homePage }, response, next);
 };
 
 module.exports = { config, serverSetup, seed };
