+++ test/config.test.js
@@ -1,5 +1,5 @@
 const test = require('ava');
-const config = require('../src/config.default.js');
+const config = require('../src/config');
 
 test('config can be parsed', (t) => {
   t.notThrows(() => {
