+++ test/delete.test.js
@@ -1,11 +1,13 @@
+// @ts-nocheck
 const test = require('ava');
 const request = require('supertest');
 const sinon = require('sinon');
-const StorageProvider = require('uttori-storage-provider-json-memory');
 
-const UttoriWiki = require('../src');
+const { UttoriWiki } = require('../src');
 
 const { config, serverSetup, seed } = require('./_helpers/server.js');
+
+const response = { set: () => {}, redirect: () => {}, render: () => {} };
 
 test('deletes the document and redirects to the home page', async (t) => {
   t.plan(2);
@@ -16,16 +18,17 @@ test('deletes the document and redirects to the home page', async (t) => {
     content: '## Delete Page',
     html: '',
     updateDate: 1412921841841,
-    createDate: null,
-    tags: [],
+    createDate: undefined,
+    tags: '',
   };
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, StorageProvider }, server);
-  seed(uttori.storageProvider);
-  uttori.storageProvider.add(testDelete);
-  const response = await request(uttori.server).get('/test-delete/delete/test-key');
-  t.is(response.status, 302);
-  t.is(response.text, 'Found. Redirecting to https://fake.test');
+  const uttori = new UttoriWiki(config, server);
+  await seed(uttori);
+  // uttori.storageProvider.add(testDelete);
+  await uttori.saveValid({ params: {}, body: testDelete }, response, () => {});
+  const express_response = await request(uttori.server).get('/test-delete/delete/test-key');
+  t.is(express_response.status, 302);
+  t.is(express_response.text, 'Found. Redirecting to https://fake.test');
 });
 
 test('falls through to next when slug is missing', async (t) => {
@@ -34,7 +37,7 @@ test('falls through to next when slug is missing', async (t) => {
   const next = sinon.spy();
   const server = serverSetup();
   const uttori = new UttoriWiki(config, server);
-  await uttori.delete({ params: { key: 'test-key' } }, null, next);
+  await uttori.delete({ params: { key: 'test-key' } }, response, next);
   t.true(next.calledOnce);
 });
 
@@ -44,7 +47,7 @@ test('falls through to next when document is not found', async (t) => {
   const next = sinon.spy();
   const server = serverSetup();
   const uttori = new UttoriWiki(config, server);
-  await uttori.delete({ params: { slug: 'missing', key: 'test-key' } }, null, next);
+  await uttori.delete({ params: { slug: 'missing', key: 'test-key' } }, response, next);
   t.true(next.calledOnce);
 });
 
@@ -53,9 +56,9 @@ test('falls to 404 when miss matched key', async (t) => {
 
   const server = serverSetup();
   const uttori = new UttoriWiki(config, server);
-  const response = await request(uttori.server).get('/missing/delete/bad-key');
-  t.is(response.status, 200);
-  t.is(response.text.slice(0, 15), '<!DOCTYPE html>');
-  const title = response.text.match(/<title>(.*?)<\/title>/i);
+  const express_response = await request(uttori.server).get('/missing/delete/bad-key');
+  t.is(express_response.status, 200);
+  t.is(express_response.text.slice(0, 15), '<!DOCTYPE html>');
+  const title = express_response.text.match(/<title>(.*?)<\/title>/i);
   t.is(title[1], '404 Not Found | Wiki');
 });
