+++ test/edit.test.js
@@ -1,9 +1,8 @@
 const test = require('ava');
 const request = require('supertest');
 const sinon = require('sinon');
-const StorageProvider = require('uttori-storage-provider-json-memory');
 
-const UttoriWiki = require('../src');
+const { UttoriWiki } = require('../src');
 
 const { config, serverSetup, seed } = require('./_helpers/server.js');
 
@@ -11,8 +10,8 @@ test('renders the edit page for a given slug', async (t) => {
   t.plan(3);
 
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, StorageProvider }, server);
-  seed(uttori.storageProvider);
+  const uttori = new UttoriWiki(config, server);
+  await seed(uttori);
   const response = await request(uttori.server).get('/demo-title/edit');
   t.is(response.status, 200);
   t.is(response.text.slice(0, 15), '<!DOCTYPE html>');
@@ -25,8 +24,8 @@ test('falls through to next when slug is missing', async (t) => {
 
   const next = sinon.spy();
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, StorageProvider }, server);
-  seed(uttori.storageProvider);
+  const uttori = new UttoriWiki(config, server);
+  await seed(uttori);
   await uttori.edit({ params: { slug: '' } }, null, next);
   t.true(next.calledOnce);
 });
@@ -36,8 +35,8 @@ test('falls through to next when document is missing', async (t) => {
 
   const next = sinon.spy();
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, StorageProvider }, server);
-  seed(uttori.storageProvider);
+  const uttori = new UttoriWiki(config, server);
+  await seed(uttori);
   await uttori.edit({ params: { slug: 'missing-document' } }, null, next);
   t.true(next.calledOnce);
 });
