+++ test/getSiteSections.test.js
@@ -1,7 +1,6 @@
 const test = require('ava');
-const StorageProvider = require('uttori-storage-provider-json-memory');
 
-const UttoriWiki = require('../src');
+const { UttoriWiki } = require('../src');
 
 const { config, serverSetup, seed } = require('./_helpers/server.js');
 
@@ -9,8 +8,8 @@ test('getSiteSections(count): returns the config sections with tag counts', asyn
   t.plan(1);
 
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, StorageProvider }, server);
-  seed(uttori.storageProvider);
+  const uttori = new UttoriWiki(config, server);
+  await seed(uttori);
   const sections = await uttori.getSiteSections();
   t.deepEqual(sections, [{
     description: 'Example description text.',
