+++ test/historyIndex.test.js
@@ -1,21 +1,23 @@
+// @ts-nocheck
 const test = require('ava');
 const request = require('supertest');
 const sinon = require('sinon');
-const StorageProvider = require('uttori-storage-provider-json-memory');
 
-const UttoriWiki = require('../src');
+const { UttoriWiki } = require('../src');
 
 const { config, serverSetup, seed } = require('./_helpers/server.js');
+
+const response = { set: () => {}, redirect: () => {}, render: () => {} };
 
 test('renders the requested slug history', async (t) => {
   t.plan(2);
 
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, StorageProvider }, server);
-  seed(uttori.storageProvider);
-  const response = await request(uttori.server).get('/demo-title/history');
-  t.is(response.status, 200);
-  const title = response.text.match(/<title>(.*?)<\/title>/i);
+  const uttori = new UttoriWiki(config, server);
+  await seed(uttori);
+  const express_response = await request(uttori.server).get('/demo-title/history');
+  t.is(express_response.status, 200);
+  const title = express_response.text.match(/<title>(.*?)<\/title>/i);
   t.is(title[1], 'Demo Title Revision History | Wiki');
 });
 
@@ -24,9 +26,9 @@ test('falls through to next when slug is missing', async (t) => {
 
   const next = sinon.spy();
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, StorageProvider }, server);
-  seed(uttori.storageProvider);
-  await uttori.historyIndex({ params: { slug: '' } }, null, next);
+  const uttori = new UttoriWiki(config, server);
+  await seed(uttori);
+  await uttori.historyIndex({ params: { slug: '' } }, response, next);
   t.true(next.calledOnce);
 });
 
@@ -35,8 +37,8 @@ test('falls through to next when no revision is found', async (t) => {
 
   const next = sinon.spy();
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, StorageProvider }, server);
-  seed(uttori.storageProvider);
-  await uttori.historyIndex({ params: { slug: 'missing' } }, null, next);
+  const uttori = new UttoriWiki(config, server);
+  await seed(uttori);
+  await uttori.historyIndex({ params: { slug: 'missing' } }, response, next);
   t.true(next.calledOnce);
 });
