+++ test/historyRestore.test.js
@@ -1,23 +1,24 @@
 const test = require('ava');
 const request = require('supertest');
 const sinon = require('sinon');
-const StorageProvider = require('uttori-storage-provider-json-memory');
 
-const UttoriWiki = require('../src');
+const { UttoriWiki } = require('../src');
 
 const { config, serverSetup, seed } = require('./_helpers/server.js');
+
+const response = { set: () => {}, render: () => {}, render: () => {} };
 
 test('renders the edit page for a given slug', async (t) => {
   t.plan(3);
 
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, StorageProvider }, server);
-  seed(uttori.storageProvider);
-  const history = uttori.storageProvider.getHistory('demo-title')[0];
-  const response = await request(uttori.server).get(`/demo-title/history/${history}/restore`);
-  t.is(response.status, 200);
-  t.is(response.text.slice(0, 15), '<!DOCTYPE html>');
-  const title = response.text.match(/<title>(.*?)<\/title>/i);
+  const uttori = new UttoriWiki(config, server);
+  await seed(uttori);
+  const [history] = await uttori.hooks.fetch('storage-get-history', 'demo-title', uttori);
+  const espress_response = await request(uttori.server).get(`/demo-title/history/${history[0]}/restore`);
+  t.is(espress_response.status, 200);
+  t.is(espress_response.text.slice(0, 15), '<!DOCTYPE html>');
+  const title = espress_response.text.match(/<title>(.*?)<\/title>/i);
   t.true(title[1].startsWith('Editing Demo Title Beta from Revision 1'));
 });
 
@@ -26,9 +27,9 @@ test('falls through to next when slug is missing', async (t) => {
 
   const next = sinon.spy();
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, StorageProvider }, server);
-  seed(uttori.storageProvider);
-  await uttori.historyRestore({ params: { slug: '' } }, null, next);
+  const uttori = new UttoriWiki(config, server);
+  await seed(uttori);
+  await uttori.historyRestore({ params: { slug: '' } }, response, next);
   t.true(next.calledOnce);
 });
 
@@ -37,9 +38,9 @@ test('falls through to next when revision is missing', async (t) => {
 
   const next = sinon.spy();
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, StorageProvider }, server);
-  seed(uttori.storageProvider);
-  await uttori.historyRestore({ params: { slug: 'demo-title', revision: '' } }, null, next);
+  const uttori = new UttoriWiki(config, server);
+  await seed(uttori);
+  await uttori.historyRestore({ params: { slug: 'demo-title', revision: '' } }, response, next);
   t.true(next.calledOnce);
 });
 
@@ -48,8 +49,8 @@ test('falls through to next when no revision is found', async (t) => {
 
   const next = sinon.spy();
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, StorageProvider }, server);
-  seed(uttori.storageProvider);
-  await uttori.historyRestore({ params: { slug: 'demo-title', revision: '1' } }, null, next);
+  const uttori = new UttoriWiki(config, server);
+  await seed(uttori);
+  await uttori.historyRestore({ params: { slug: 'demo-title', revision: '1' } }, response, next);
   t.true(next.calledOnce);
 });
