+++ test/home.test.js
@@ -1,20 +1,34 @@
+// @ts-nocheck
 const test = require('ava');
 const request = require('supertest');
-const StorageProvider = require('uttori-storage-provider-json-memory');
+const sinon = require('sinon');
 
-const UttoriWiki = require('../src');
+const { UttoriWiki } = require('../src');
 
 const { config, serverSetup, seed } = require('./_helpers/server.js');
 
-test('home(request, response, _next): renders', async (t) => {
+const response = { set: () => {}, redirect: () => {}, render: () => {} };
+
+test('renders', async (t) => {
   t.plan(3);
 
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, StorageProvider }, server);
-  seed(uttori.storageProvider);
-  const response = await request(uttori.server).get('/');
-  t.is(response.status, 200);
-  t.is(response.text.slice(0, 15), '<!DOCTYPE html>');
-  const title = response.text.match(/<title>(.*?)<\/title>/i);
+  const uttori = new UttoriWiki(config, server);
+  await seed(uttori);
+  const express_response = await request(uttori.server).get('/');
+  t.is(express_response.status, 200);
+  t.is(express_response.text.slice(0, 15), '<!DOCTYPE html>');
+  const title = express_response.text.match(/<title>(.*?)<\/title>/i);
   t.is(title[1], 'Home | Wiki');
+});
+
+test('falls through to next when home document is missing', async (t) => {
+  t.plan(1);
+
+  const next = sinon.spy();
+  const server = serverSetup();
+  const uttori = new UttoriWiki({ ...config, home_page: '' }, server);
+  await seed(uttori);
+  await uttori.home({ params: { slug: '' } }, response, next);
+  t.true(next.calledOnce);
 });
