+++ test/index.test.js
@@ -1,6 +1,6 @@
 const test = require('ava');
 
-const UttoriWiki = require('../src');
+const { UttoriWiki } = require('../src');
 
 const { config, serverSetup } = require('./_helpers/server.js');
