+++ test/new.test.js
@@ -1,7 +1,7 @@
 const test = require('ava');
 const request = require('supertest');
 
-const UttoriWiki = require('../src');
+const { UttoriWiki } = require('../src');
 
 const { config, serverSetup, seed } = require('./_helpers/server.js');
 
@@ -10,7 +10,7 @@ test('new(request, response, _next): renders', async (t) => {
 
   const server = serverSetup();
   const uttori = new UttoriWiki(config, server);
-  seed(uttori.storageProvider);
+  await seed(uttori);
   const response = await request(uttori.server).get('/new');
   t.is(response.status, 200);
   t.is(response.text.slice(0, 15), '<!DOCTYPE html>');
