+++ test/registerPlugins.test.js
@@ -1,9 +1,6 @@
 const test = require('ava');
 
-const StorageProvider = require('uttori-storage-provider-json-memory');
-const SearchProvider = require('./__stubs/SearchProvider.js');
-
-const UttoriWiki = require('../src');
+const { UttoriWiki } = require('../src');
 
 const { serverSetup } = require('./_helpers/server.js');
 
@@ -11,24 +8,9 @@ test('registerPlugins(config): does not fail when plugins is broken', (t) => {
   t.notThrows(() => {
     const server = serverSetup();
     const _uttori = new UttoriWiki({
-      StorageProvider,
-      SearchProvider,
       theme_dir: '/tmp',
       public_dir: '/tmp',
       plugins: {},
-    }, server);
-  });
-});
-
-test('registerPlugins(config): does not load plugins that do not begin with uttori-plugin-', (t) => {
-  t.notThrows(() => {
-    const server = serverSetup();
-    const _uttori = new UttoriWiki({
-      StorageProvider,
-      SearchProvider,
-      theme_dir: '/tmp',
-      public_dir: '/tmp',
-      plugins: ['fake'],
     }, server);
   });
 });
