+++ test/tag.test.js
@@ -1,8 +1,7 @@
 const test = require('ava');
 const request = require('supertest');
-const StorageProvider = require('uttori-storage-provider-json-memory');
 
-const UttoriWiki = require('../src');
+const { UttoriWiki } = require('../src');
 
 const { config, serverSetup, seed } = require('./_helpers/server.js');
 
@@ -10,8 +9,8 @@ test('tagsIndex(request, response, next): renders that tag index page', async (t
   t.plan(3);
 
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, StorageProvider }, server);
-  seed(uttori.storageProvider);
+  const uttori = new UttoriWiki(config, server);
+  await seed(uttori);
   const response = await request(uttori.server).get('/tags');
   t.is(response.status, 200);
   t.is(response.text.slice(0, 15), '<!DOCTYPE html>');
@@ -23,8 +22,8 @@ test('tag(request, response, next): renders that tag page for a given tag', asyn
   t.plan(3);
 
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, StorageProvider }, server);
-  seed(uttori.storageProvider);
+  const uttori = new UttoriWiki(config, server);
+  await seed(uttori);
   const response = await request(uttori.server).get('/tags/Cool');
   t.is(response.status, 200);
   t.is(response.text.slice(0, 15), '<!DOCTYPE html>');
@@ -36,8 +35,8 @@ test('tag(request, response, next): falls through to next when tag is missing', 
   t.plan(3);
 
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, StorageProvider }, server);
-  seed(uttori.storageProvider);
+  const uttori = new UttoriWiki(config, server);
+  await seed(uttori);
   const response = await request(uttori.server).get('/tags/_');
   t.is(response.status, 200);
   t.is(response.text.slice(0, 15), '<!DOCTYPE html>');
